+++ MercadoPagoSDK/MercadoPagoSDK/Flows/OneTap/UI/PXOneTapHeaderProtocol.swift
@@ -7,11 +7,6 @@
 
 import Foundation
 
-protocol PXOneTapSummaryProtocol: NSObjectProtocol {
-    func didTapCharges()
-    func didTapDiscount()
-}
-
 protocol PXOneTapHeaderProtocol: PXOneTapSummaryProtocol {
     func didTapMerchantHeader()
     func splitPaymentSwitchChangedValue(isOn: Bool, isUserSelection: Bool)
