+++ MercadoPagoSDK/MercadoPagoSDK/Flows/OneTap/UI/PXOneTapHeaderViewModel.swift
@@ -7,14 +7,7 @@
 
 import UIKit
 
-typealias OneTapSummaryRowHandler = () -> Void
-typealias OneTapHeaderSummaryData = (title: String, value: String, highlightedColor: UIColor, alpha: CGFloat, isTotal: Bool, image: UIImage?, type: OneTapSummaryRowType?)
-
-enum OneTapSummaryRowType {
-    case discount
-    case charges
-    case generic
-}
+typealias OneTapHeaderSummaryData = (title: String, value: String, highlightedColor: UIColor, alpha: CGFloat, isTotal: Bool, image: UIImage?, type: PXOneTapSummaryRowView.RowType?)
 
 class PXOneTapHeaderViewModel {
     let icon: UIImage
