+++ MercadoPagoSDK/MercadoPagoSDK/Flows/OneTap/UI/PXOneTapSummaryProtocol.swift
@@ -1,0 +1,13 @@
+//
+//  PXOneTapSummaryProtocol.swift
+//  MercadoPagoSDK
+//
+//  Created by Federico Bustos Fierro on 22/05/2019.
+//
+
+import Foundation
+
+protocol PXOneTapSummaryProtocol: NSObjectProtocol {
+    func didTapCharges()
+    func didTapDiscount()
+}
